Dump stack traces of all threads on test failure due to AssertionError
Closes #12425
